feat: new API for all components
New API as described in #14 implemented. Refer to issue for details
and reasoning

BREAKING CHANGE: Labels and *all* text configs/props now go in their
own new components

resolves #14 fixes #17 fixes #15
