+++ .storybook/stories/Alignment.stories.tsx
@@ -3,8 +3,9 @@ import { storiesOf } from "@storybook/react";
 import "./index.css";
 import { Scene } from "./common";
 
-import { Input } from "../../src/components/Input";
-import { Textarea } from "../../src/components/Textarea";
+import { Input, Textarea, Label } from "../../src";
+import { Text as InputText } from "../../src/Input";
+import { Text as TextareaText } from "../../src/Textarea";
 import { Vector2 } from "@react-three/fiber";
 
 const stories = storiesOf("Alignment", module);
@@ -15,31 +16,30 @@ stories.add("Vertical Alignment Input/Textarea", () => {
 
   return (
     <Scene lightColor="blue">
-      <mesh position-y={fontSize / 2 + padding[1]}>
-        <planeGeometry args={[2, 0.001]} />
-        <meshBasicMaterial color="red" />
-      </mesh>
+      <group position-z={0.25}>
+        <mesh position-y={fontSize / 2 + padding[1]}>
+          <planeGeometry args={[2, 0.0025]} />
+          <meshBasicMaterial color="red" />
+        </mesh>
 
-      <mesh position-y={-fontSize / 2}>
-        <planeGeometry args={[2, 0.001]} />
-        <meshBasicMaterial color="red" />
-      </mesh>
+        <mesh position-y={-fontSize / 2}>
+          <planeGeometry args={[2, 0.0025]} />
+          <meshBasicMaterial color="red" />
+        </mesh>
 
-      <Input
-        label="Text Input:"
-        width={1}
-        position-x={-0.51}
-        padding={padding}
-        textProps={{ fontSize: fontSize }}
-      />
-      <Textarea
-        label="Text Input:"
-        width={1}
-        rows={6}
-        position-x={0.51}
-        padding={padding}
-        textProps={{ fontSize: fontSize }}
-      />
+        <group position-x={-0.51}>
+          <Label text="label" />
+          <Input width={1} padding={padding}>
+            <InputText fontSize={fontSize} />
+          </Input>
+        </group>
+        <group position-x={0.51}>
+          <Label text="label" />
+          <Textarea width={1} rows={6} padding={padding}>
+            <TextareaText fontSize={fontSize} />
+          </Textarea>
+        </group>
+      </group>
     </Scene>
   );
 });
