+++ src/Textarea/index.ts
@@ -1,1 +1,2 @@
-export * from "./Textarea";
+export { Textarea } from "./Textarea";
+export { Text } from "./Text";
