+++ src/index.ts
@@ -2,3 +2,4 @@ export { Input } from "./Input";
 export { Label } from "./Label";
 export { Textarea } from "./Textarea";
 export { Form } from "./Form";
+export { Submit } from "./Submit";
